# backend/src/openapi-generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MicroPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MicroPay — instant micropayments on the Stellar network.\n\n## Rate Limiting\n\nAll endpoints are rate-limited. Three limiters apply:\n\n| Limiter | Window | Limit | Routes |\n|---------|--------|-------|--------|\n| Global | 15 minutes | 100 req/IP | All routes |\n| Strict | 1 minute | 20 req/IP | `/api/analytics/*`, `/api/tips/*`, `/api/webhooks/*`, `/federation`, `/api/accounts/register` |\n| Payment | 1 minute | 10 req/IP | `/api/payments/*`, `/api/turrets/*` |\n\nEvery response includes the following headers so clients can implement back-off:\n\n| Header | Description |\n|--------|-------------|\n| `RateLimit-Limit` | Maximum requests allowed in the current window |\n| `RateLimit-Remaining` | Requests remaining before the limit is reached |\n| `RateLimit-Reset` | Seconds until the window resets |\n\nWhen the limit is exceeded the server returns **HTTP 429** with `{ \"error\": \"Too many requests, please try again later.\" }`. Clients should read `RateLimit-Remaining` on each response and add exponential back-off when the value approaches 0.",
    "contact": {
      "name": "Stellar MicroPay",
      "url": "https://github.com/Emmy123222/Stellar-MicroPay"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/turrets": {
      "get": {
        "tags": [
          "Turrets"
        ],
        "summary": "List txFunction deployments",
        "description": "Returns all deployments. Filter by owner using `ownerPublicKey` query parameter.",
        "parameters": [
          {
            "name": "ownerPublicKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            },
            "description": "Filter deployments by owner Stellar public key"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "ownerPublicKey": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "dca",
                              "stop_loss",
                              "escrow_release"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "paused",
                              "completed"
                            ]
                          },
                          "config": {
                            "type": "object"
                          },
                          "deploymentHash": {
                            "type": "string"
                          },
                          "signedChallengeXDR": {
                            "type": "string",
                            "description": "Owner-signed challenge transaction XDR used to deploy"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAtMs": {
                            "type": "integer",
                            "description": "Creation time in epoch milliseconds"
                          },
                          "nextRunAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastExecutedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastCheckedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "lastObservedPriceUsd": {
                            "type": "number",
                            "nullable": true
                          },
                          "lastError": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ownerPublicKey format"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/turrets/challenge": {
      "post": {
        "tags": [
          "Turrets"
        ],
        "summary": "Create a txFunction signing challenge",
        "description": "Returns a ManageData transaction XDR that the user must sign with their Stellar keypair to prove ownership.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ownerPublicKey",
                  "type",
                  "config"
                ],
                "properties": {
                  "ownerPublicKey": {
                    "type": "string",
                    "pattern": "^G[A-Z0-9]{55}$",
                    "description": "Stellar public key of the txFunction owner"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "dca",
                      "stop_loss",
                      "escrow_release"
                    ],
                    "description": "Type of automated txFunction"
                  },
                  "config": {
                    "type": "object",
                    "description": "Type-specific configuration — see DcaConfig, StopLossConfig, or EscrowReleaseConfig"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge XDR and deployment hash",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "challengeXDR": {
                          "type": "string",
                          "description": "Base64-encoded ManageData transaction XDR the owner must sign"
                        },
                        "deploymentHash": {
                          "type": "string",
                          "description": "SHA-256 hash of the normalised config, included in the challenge"
                        },
                        "normalizedConfig": {
                          "type": "object"
                        },
                        "networkPassphrase": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body (bad public key, unknown type, invalid config)"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/turrets/deploy": {
      "post": {
        "tags": [
          "Turrets"
        ],
        "summary": "Deploy a signed txFunction",
        "description": "Verifies the signed challenge and registers the txFunction. The runner begins evaluating the deployment immediately.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ownerPublicKey",
                  "type",
                  "config",
                  "deploymentHash",
                  "signedChallengeXDR"
                ],
                "properties": {
                  "ownerPublicKey": {
                    "type": "string",
                    "pattern": "^G[A-Z0-9]{55}$"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "dca",
                      "stop_loss",
                      "escrow_release"
                    ]
                  },
                  "config": {
                    "type": "object"
                  },
                  "deploymentHash": {
                    "type": "string"
                  },
                  "signedChallengeXDR": {
                    "type": "string",
                    "description": "The challenge XDR signed by the owner's Freighter (or Ledger) wallet"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "ownerPublicKey": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "dca",
                            "stop_loss",
                            "escrow_release"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "paused",
                            "completed"
                          ]
                        },
                        "config": {
                          "type": "object"
                        },
                        "deploymentHash": {
                          "type": "string"
                        },
                        "signedChallengeXDR": {
                          "type": "string",
                          "description": "Owner-signed challenge transaction XDR used to deploy"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAtMs": {
                          "type": "integer",
                          "description": "Creation time in epoch milliseconds"
                        },
                        "nextRunAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastExecutedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastCheckedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastObservedPriceUsd": {
                          "type": "number",
                          "nullable": true
                        },
                        "lastError": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Config hash mismatch or invalid asset"
          },
          "401": {
            "description": "Signed challenge was not signed by the owner"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/turrets/{id}": {
      "get": {
        "tags": [
          "Turrets"
        ],
        "summary": "Get a single txFunction deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "ownerPublicKey": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "dca",
                            "stop_loss",
                            "escrow_release"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "paused",
                            "completed"
                          ]
                        },
                        "config": {
                          "type": "object"
                        },
                        "deploymentHash": {
                          "type": "string"
                        },
                        "signedChallengeXDR": {
                          "type": "string",
                          "description": "Owner-signed challenge transaction XDR used to deploy"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAtMs": {
                          "type": "integer",
                          "description": "Creation time in epoch milliseconds"
                        },
                        "nextRunAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastExecutedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastCheckedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "lastObservedPriceUsd": {
                          "type": "number",
                          "nullable": true
                        },
                        "lastError": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  }
}